chore : auth디렉토리 생성 후 클래스위치 변경
