+++ src/main/java/com/zerobase/nsbackend/auth/config/AppConfig.java
@@ -1,4 +1,4 @@
-package com.zerobase.nsbackend.member.config;
+package com.zerobase.nsbackend.auth.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
