+++ src/main/java/com/zerobase/nsbackend/auth/security/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.zerobase.nsbackend.member.security;
+package com.zerobase.nsbackend.auth.security;
 
 
 import java.io.IOException;
