+++ src/main/java/com/zerobase/nsbackend/auth/service/AuthService.java
@@ -1,10 +1,10 @@
-package com.zerobase.nsbackend.member.service;
+package com.zerobase.nsbackend.auth.service;
 
 import com.zerobase.nsbackend.global.exceptionHandle.ErrorCode;
 import com.zerobase.nsbackend.member.domain.Member;
 import com.zerobase.nsbackend.member.domain.repository.MemberRepository;
-import com.zerobase.nsbackend.member.dto.Auth.SignIn;
-import com.zerobase.nsbackend.member.dto.Auth.SignUp;
+import com.zerobase.nsbackend.auth.dto.Auth.SignIn;
+import com.zerobase.nsbackend.auth.dto.Auth.SignUp;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
